+++ proc/libs/text_lib/texlib.cpp
@@ -35,7 +35,7 @@ void TextDtor (Text* text)
 
 //flexxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
 
-int CountSize (FILE* file)
+size_t CountSize (FILE *file)
 {   
     assert (file);
