+++ proc/libs/text_lib/texlib.h
@@ -73,7 +73,7 @@ void TextDtor (Text *text);
 
 // flexxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
 
-int CountSize (FILE *file);
+size_t CountSize (FILE *file);
 
 // flexxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
